+++ dbms/db2/src/main/java/is/codion/dbms/db2/Db2DatabaseFactory.java
@@ -31,7 +31,7 @@ public final class Db2DatabaseFactory implements DatabaseFactory {
 	private static final String DRIVER_PACKAGE = "com.ibm.db2.jcc";
 
 	@Override
-	public boolean driverCompatible(String driverClassName) {
+	public boolean supports(String driverClassName) {
 		return requireNonNull(driverClassName).startsWith(DRIVER_PACKAGE);
 	}