+++ dbms/sqlite/src/main/java/is/codion/dbms/sqlite/SQLiteDatabaseFactory.java
@@ -31,7 +31,7 @@ public final class SQLiteDatabaseFactory implements DatabaseFactory {
 	private static final String DRIVER_PACKAGE = "org.sqlite";
 
 	@Override
-	public boolean driverCompatible(String driverClassName) {
+	public boolean supports(String driverClassName) {
 		return requireNonNull(driverClassName).startsWith(DRIVER_PACKAGE);
 	}